requirements.txt
webob
routes
pyyaml
requests
sleekxmpp    # dev only
mock         # dev only
jsonrpclib   # dev only
